pyproject.toml
[tool.poetry]
name = "dax-hoc"
version = "0.2.0"
description = "Short-term earnings news trader for DAX/MDAX/SDAX companies"
authors = ["Martin Bierey"]
packages = [
  {include = "config.py", from = "src"},
  {include = "db", from = "src"},
  {include = "universe", from = "src"},
  {include = "eqs", from = "src"},
  {include = "earnings", from = "src"},
  {include = "expectations", from = "src"},
  {include = "reaction", from = "src"},
  {include = "llm", from = "src"},
  {include = "ensemble", from = "src"},
  {include = "pipeline", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.28"
sqlalchemy = "^2.0"
python-dotenv = "^1.0"
beautifulsoup4 = "^4.12"
pandas = "^3.0"
anthropic = "^0.50"
openai = "^1.0"
jupyterlab = "^4.5"
google-genai = "^1.72.0"
yfinance = "^1.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11"
pytest = "^9.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
